Rename functions file as it is specific to users
