+++ packages/nodes-base/nodes/Okta/UserDescription.ts
@@ -1,5 +1,5 @@
 import type { INodeProperties } from 'n8n-workflow';
-import { simplifyGetAllResponse, simplifyGetResponse } from './GenericFunctions';
+import { simplifyGetAllResponse, simplifyGetResponse } from './UserFunctions';
 const BASE_API_URL = '/api/v1/users/';
 export const userOperations: INodeProperties[] = [
 	{
